Refactors moved stores. Changes the way they are registered.
