--- src/themes/default/store/claims/store/index.ts
@@ -1,9 +1,0 @@
-import { Module } from 'vuex'
-import actions from './actions'
-import RootState from '@vue-storefront/core/types/RootState'
-import ClaimsState from '../types/ClaimsState'
-
-export const module: Module<ClaimsState, RootState> = {
-  namespaced: true,
-  actions
-}
