--- src/themes/default/store/claims/types/ClaimsState.ts
@@ -1,3 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export default interface ClaimsState {
-}
