+++ src/themes/default/store/homepage.ts
@@ -1,0 +1,15 @@
+// import { createModule } from '@vue-storefront/core/lib/module'
+
+// const KEY = 'homepage'
+export const store = {
+  namespaced: true,
+  state: {
+    new_collection: []
+  }
+}
+
+
+// export const Homepage = createModule({
+//   key: KEY,
+//   store: { modules: [{ key: KEY, module: store }] }
+// })
