--- src/themes/default/store/homepage/index.ts
@@ -1,13 +1,0 @@
-import { createModule } from '@vue-storefront/core/lib/module'
-
-const KEY = 'homepage'
-const store = {
-  namespaced: true,
-  state: {
-    new_collection: []
-  }
-}
-export const Homepage = createModule({
-  key: KEY,
-  store: { modules: [{ key: KEY, module: store }] }
-})
