--- src/themes/default/store/promoted-offers/types/PromotedOffersState.ts
@@ -1,8 +1,0 @@
-export default interface PromotedOffersState {
-  banners: {
-    mainBanners: any[],
-    smallBanners: any[],
-    productBanners: any[]
-  },
-  headImage: Record<string, any>
-}
